refactor: Moved workflow decorator logic to a separate file Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
