+++ typescript/vrotsc/src/compiler/transformer/fileTransformers/workflow/workflow.ts
@@ -1,21 +1,22 @@
 import * as ts from "typescript";
 
-import { FileDescriptor, FileTransformationContext, FileType, DiagnosticCategory, XmlNode, XmlElement } from "../../../../types";
+import { FileDescriptor, FileTransformationContext, FileType, XmlNode, XmlElement } from "../../../../types";
 import { system } from "../../../../system/system";
 import { printElementInfo } from "../../../elementInfo";
 import { transformSourceFile } from "../../scripts/scripts";
 import { collectFactsBefore } from "../../metaTransformers/facts";
 import { transformShimsBefore, transformShims } from "../../codeTransformers/shims";
-import { addHeaderComment, printSourceFile } from "../../helpers/source";
+import { printSourceFile } from "../../helpers/source";
 import { generateElementId } from "../../../../utilities/utilities";
-import { getPropertyName, getIdentifierTextOrNull, getDecoratorNames, getDecoratorName } from "../../helpers/node";
-import { getVroType } from "../../helpers/vro";
+import { getPropertyName } from "../../helpers/node";
 import { StringBuilderClass } from "../../../../utilities/stringBuilder";
-import { WorkflowDescriptor, WorkflowItemDescriptor, WorkflowParameterType, WorkflowParameter, PolyglotDescriptor, WorkflowMethodType } from "../../../../decorators";
+import { WorkflowDescriptor, WorkflowItemDescriptor, WorkflowParameterType, WorkflowParameter, PolyglotDescriptor } from "../../../../decorators";
 import { remediateTypeScript } from "../../codeTransformers/remediate";
 import { transformModuleSystem } from "../../codeTransformers/modules";
 import { buildPolyglotInfo, printPolyglotCode } from "./polyglot";
 import { prepareHeaderEmitter } from "../../codeTransformers/header";
+import { createWorkflowItemPrologueStatements } from "../../codeTransformers/prologueStatements";
+import { buildWorkflowDecorators, registerMethodArgumentDecorators } from "./decorators";
 
 const xmldoc: typeof import("xmldoc") = require("xmldoc");
 
@@ -97,7 +98,12 @@ export function getWorkflowTransformer(file: FileDescriptor, context: FileTransf
 		};
 		const decorators = ts.getDecorators(classNode);
 		if (decorators && decorators.length) {
-			buildWorkflowDecorators(workflowInfo, classNode);
+			buildWorkflowDecorators(
+				workflowInfo,
+				classNode,
+				context,
+				sourceFile
+			);
 		}
 
 		classNode.members
@@ -169,264 +175,6 @@ export function getWorkflowTransformer(file: FileDescriptor, context: FileTransf
 		workflowInfo.items.push(itemInfo);
 	}
 
-	/**
-	 * Fetches details from the decorators for the methods and adds the information to the Descriptors
-	 */
-	function registerMethodDecorators(methodNode: ts.MethodDeclaration, workflowInfo: WorkflowDescriptor, itemInfo: WorkflowItemDescriptor) {
-		const decorators = ts.getDecorators(methodNode);
-
-		if (decorators && decorators.length !== 1) {
-			throw new Error(`Method '${itemInfo.name}' should have exactly one decorator.`);
-		}
-
-		const decoratorName = getDecoratorName(decorators[0]);
-
-		switch (decoratorName) {
-			case WorkflowMethodType.Item:
-
-				break;
-		}
-	}
-
-	function registerMethodArgumentDecorators(methodNode: ts.MethodDeclaration, workflowInfo: WorkflowDescriptor, itemInfo: WorkflowItemDescriptor) {
-		if (methodNode.parameters.length) {
-			methodNode.parameters.forEach(paramNode => {
-				const name = getIdentifierTextOrNull(paramNode.name);
-				if (name) {
-					let parameterType = WorkflowParameterType.Default;
-					const decorators = ts.getDecorators(methodNode);
-					// Adds state of what decorators are present
-					getDecoratorNames(decorators).forEach(decoratorName => {
-						switch (decoratorName || "In") {
-							case "In":
-								parameterType |= WorkflowParameterType.Input;
-								break;
-							case "Out":
-								parameterType |= WorkflowParameterType.Output;
-								break;
-							default:
-								throw new Error(`Decorator '${decoratorName} is not supported'`);
-						}
-					});
-
-					if (parameterType === WorkflowParameterType.Default) {
-						parameterType = WorkflowParameterType.Input;
-					}
-
-					if (parameterType & WorkflowParameterType.Input) {
-						itemInfo.input.push(name);
-					}
-
-					if (parameterType & WorkflowParameterType.Output) {
-						itemInfo.output.push(name);
-					}
-
-					buildWorkflowParameter(workflowInfo, paramNode, parameterType);
-				}
-			});
-		}
-	}
-
-	function createWorkflowItemPrologueStatements(methodNode: ts.MethodDeclaration): ts.Statement[] {
-		const statements: ts.Statement[] = [];
-
-		if (methodNode.parameters.length) {
-			const variableDeclarations: ts.VariableDeclaration[] = [];
-			methodNode.parameters.forEach(paramNode => {
-				const paramName = (<ts.Identifier>paramNode.name).text;
-				variableDeclarations.push(ts.factory.createVariableDeclaration(
-					paramName,
-					undefined,
-					paramNode.type,
-                        /* initializer */ undefined
-				));
-			});
-
-			if (variableDeclarations.length) {
-				statements.push(ts.factory.createVariableStatement(
-					[ts.factory.createModifier(ts.SyntaxKind.DeclareKeyword)],
-					variableDeclarations));
-			}
-		}
-
-		return statements;
-	}
-
-	function buildWorkflowParameter(workflowInfo: WorkflowDescriptor, paramNode: ts.ParameterDeclaration, parameterType: WorkflowParameterType): void {
-		const name = (<ts.Identifier>paramNode.name).text;
-		let parameter = workflowInfo.parameters.find(p => p.name === name);
-		if (!parameter) {
-			parameter = {
-				name: name,
-				type: getVroType(paramNode.type),
-				parameterType: WorkflowParameterType.Default,
-				required: !paramNode.questionToken
-			};
-			workflowInfo.parameters.push(parameter);
-		}
-
-		parameter.parameterType |= parameterType;
-
-		if (parameter.type == null) {
-			parameter.type = getVroType(paramNode.type);
-		}
-	}
-
-	function buildWorkflowDecorators(workflowInfo: WorkflowDescriptor, classNode: ts.ClassDeclaration): void {
-		const decorators = ts.getDecorators(classNode);
-		decorators
-			.filter(decoratorNode => {
-				const callExpNode = decoratorNode.expression as ts.CallExpression;
-				if (callExpNode && callExpNode.expression.kind === ts.SyntaxKind.Identifier) {
-					return (<ts.Identifier>callExpNode.expression).text === "Workflow";
-				}
-			})
-			.forEach(decoratorNode => {
-				buildWorkflowDecorator(workflowInfo, <ts.CallExpression>decoratorNode.expression);
-			});
-	}
-
-	function buildWorkflowDecorator(workflowInfo: WorkflowDescriptor, decoratorCallExp: ts.CallExpression): void {
-		const objLiteralNode = decoratorCallExp.arguments[0] as ts.ObjectLiteralExpression;
-		if (objLiteralNode) {
-			objLiteralNode.properties.forEach((property: ts.PropertyAssignment) => {
-				const propName = getPropertyName(property.name);
-				switch (propName) {
-					case "id":
-						workflowInfo.id = (<ts.StringLiteral>property.initializer).text;
-						break;
-					case "name":
-						workflowInfo.name = (<ts.StringLiteral>property.initializer).text;
-						break;
-					case "path":
-						workflowInfo.path = (<ts.StringLiteral>property.initializer).text;
-						break;
-					case "version":
-						workflowInfo.version = (<ts.StringLiteral>(property.initializer)).text;
-						break;
-					case "presentation":
-						workflowInfo.presentation = (<ts.StringLiteral>property.initializer).text;
-						break;
-					case "description":
-						workflowInfo.description = (<ts.StringLiteral>property.initializer).text;
-						break;
-					case "input":
-						buildWorkflowDecoratorParameters(workflowInfo.parameters, <ts.ObjectLiteralExpression>property.initializer, WorkflowParameterType.Input);
-						break;
-					case "output":
-						buildWorkflowDecoratorParameters(workflowInfo.parameters, <ts.ObjectLiteralExpression>property.initializer, WorkflowParameterType.Output);
-						break;
-					case "attributes":
-						buildWorkflowDecoratorParameters(workflowInfo.parameters, <ts.ObjectLiteralExpression>property.initializer, WorkflowParameterType.Default, true);
-						break;
-					default:
-						throw new Error(`Workflow attribute '${propName}' is not suported.`);
-				}
-			});
-		}
-	}
-
-	function buildWorkflowDecoratorParameters(parameters: WorkflowParameter[], objLiteralNode: ts.ObjectLiteralExpression,
-		parameterType: WorkflowParameterType, isAttribute?: boolean): void {
-		objLiteralNode.properties.forEach((property: ts.PropertyAssignment) => {
-			const name = getPropertyName(property.name);
-			const objectLiteralNode = <ts.ObjectLiteralExpression>property.initializer;
-			if (objectLiteralNode) {
-				const parameter = <WorkflowParameter>{
-					name: name,
-					parameterType: parameterType,
-					isAttribute: isAttribute,
-				};
-
-				objectLiteralNode.properties.forEach((property: ts.PropertyAssignment) => {
-					const propName = getPropertyName(property.name);
-					switch (propName) {
-						case "type":
-							parameter.type = (<ts.StringLiteral>property.initializer).text;
-							break;
-						case "title":
-							parameter.title = (<ts.StringLiteral>property.initializer).text;
-							break;
-						case "required":
-							parameter.required = property.initializer.kind === ts.SyntaxKind.TrueKeyword;
-							break;
-						case "description":
-							parameter.description = (<ts.StringLiteral>property.initializer).text;
-							break;
-						case "multiLine":
-							parameter.multiLine = property.initializer.kind === ts.SyntaxKind.TrueKeyword;
-							break;
-						case "hidden":
-							parameter.hidden = property.initializer.kind === ts.SyntaxKind.TrueKeyword;
-							break;
-						case "maxStringLength":
-							parameter.maxStringLength = parseInt((<ts.NumericLiteral>property.initializer).text);
-							break;
-						case "minStringLength":
-							parameter.minStringLength = parseInt((<ts.NumericLiteral>property.initializer).text);
-							break;
-						case "numberFormat":
-							parameter.numberFormat = (<ts.StringLiteral>property.initializer).text;
-							break;
-						case "defaultValue":
-						case "value":
-							{
-								parameter.defaultValue = getWorkflowParamValue(property.initializer);
-								if (parameter.defaultValue === undefined) {
-									context.diagnostics.addAtNode(
-										sourceFile,
-										property.initializer,
-										`Workflow parameter default value should be of type string, number or boolean.`,
-										DiagnosticCategory.Error);
-								}
-							}
-							break;
-						case "availableValues":
-							{
-								parameter.availableValues = (<ts.ArrayLiteralExpression>property.initializer).elements.map(getWorkflowParamValue);
-								if (parameter.availableValues.some(v => v === undefined)) {
-									context.diagnostics.addAtNode(
-										sourceFile,
-										property.initializer,
-										`Workflow parameter available values should be of type string, number or boolean.`,
-										DiagnosticCategory.Error);
-									parameter.availableValues = undefined;
-								}
-							}
-							break;
-						case "bind":
-							{
-								parameter.bind = property.initializer.kind === ts.SyntaxKind.TrueKeyword;
-								break;
-							}
-						default:
-							throw new Error(`Workflow parameter attribute '${propName}' is not suported.`);
-					}
-				});
-
-				parameters.push(parameter);
-			}
-		});
-	}
-
-	function getWorkflowParamValue(node: ts.Node): string {
-		switch (node.kind) {
-			case ts.SyntaxKind.StringLiteral:
-				return (<ts.StringLiteral>node).text;
-			case ts.SyntaxKind.NumericLiteral:
-				{
-					let value = (<ts.NumericLiteral>node).text;
-					if (value.indexOf(".") < 0) {
-						value += ".0";
-					}
-					return value;
-				}
-			case ts.SyntaxKind.TrueKeyword:
-				return "true";
-			case ts.SyntaxKind.FalseKeyword:
-				return "false";
-		}
-	}
 
 	function mergeWorkflowXml(workflow: WorkflowDescriptor, xmlTemplate: string): string {
 		console.log(`Merging ${workflow.name}`);
